# Add support for empty replica pages