--- a/gms/feature_service.hh
+++ b/gms/feature_service.hh
@@ -103,6 +103,8 @@ public:
     // sure the whole cluster supports the new page_size field and we can safely
     // send it to replicas.
     gms::feature separate_page_size_and_safety_limit { *this, "SEPARATE_PAGE_SIZE_AND_SAFETY_LIMIT"sv };
+    // Replica is allowed to send back empty pages to coordinator on queries.
+    gms::feature empty_replica_pages { *this, "EMPTY_REPLICA_PAGES"sv };
     gms::feature supports_raft_cluster_mgmt { *this, "SUPPORTS_RAFT_CLUSTER_MANAGEMENT"sv };
     gms::feature uses_raft_cluster_mgmt { *this, "USES_RAFT_CLUSTER_MANAGEMENT"sv };
     gms::feature tombstone_gc_options { *this, "TOMBSTONE_GC_OPTIONS"sv };