--- a/multishard_mutation_query.cc
+++ b/multishard_mutation_query.cc
@@ -831,9 +831,9 @@ private:
 
 public:
     data_query_result_builder(const schema& s, const query::partition_slice& slice, query::result_options opts,
-            query::result_memory_accounter&& accounter, const compact_for_query_state_v2& compaction_state)
+            query::result_memory_accounter&& accounter, const compact_for_query_state_v2& compaction_state, uint64_t tombstone_limit)
         : _compaction_state(compaction_state)
-        , _res_builder(std::make_unique<query::result::builder>(slice, opts, std::move(accounter)))
+        , _res_builder(std::make_unique<query::result::builder>(slice, opts, std::move(accounter), tombstone_limit))
         , _builder(s, *_res_builder) { }
 
     void consume_new_partition(const dht::decorated_key& dk) { _builder.consume_new_partition(dk); }
@@ -880,6 +880,6 @@ future<std::tuple<foreign_ptr<lw_shared_ptr<query::result>>, cache_temperature>>
 
     return do_query_on_all_shards<data_query_result_builder>(db, query_schema, cmd, ranges, std::move(trace_state), timeout,
             [table_schema, &cmd, opts] (query::result_memory_accounter&& accounter, const compact_for_query_state_v2& compaction_state) {
-        return data_query_result_builder(*table_schema, cmd.slice, opts, std::move(accounter), compaction_state);
+        return data_query_result_builder(*table_schema, cmd.slice, opts, std::move(accounter), compaction_state, cmd.tombstone_limit);
     });
 }