--- a/query.cc
+++ b/query.cc
@@ -410,11 +410,9 @@ foreign_ptr<lw_shared_ptr<query::result>> result_merger::get() {
         });
         if (r->is_short_read()) {
             is_short_read = short_read::yes;
-            last_position.reset();
             break;
         }
         if (row_count >= _max_rows || partition_count >= _max_partitions) {
-            last_position.reset();
             break;
         }
     }