… cql3/statements/select_statement.cc
… idl/storage_proxy.idl.hh
… redis/query_utils.cc
… replica/database.hh
… replica/table.cc
… service/pager/query_pagers.cc
… service/storage_proxy.cc
… service/storage_proxy.hh
… test/boost/database_test.cc
… test/boost/multishard_mutation_query_test.cc
… test/boost/mutation_query_test.cc
… test/boost/per_partition_rate_limit_test.cc
… test/boost/querier_cache_test.cc
… test/cql-pytest/test_tombstone_limit.py
… test/cql-pytest/util.py
… thrift/handler.cc