@avikivity 2022-07-18
> TODO:
> 
> * Handle mismatching continue positions on the coordinator: if the digest matches, just pick the smallest one as the continue position.

Sounds simple.

> * Multi-node tests, with various mismatching amount of tombstones on different nodes.
> * [optimization] allow querier lookup to succeed if continue position is < then querier's position, if we know that the range `[$continue_pos, $querier_pos]` contains only dead data.


We should only do this if it's a problem in practice (and we have counters to see if it's a problem or not).

> 
> Questions:
> 
> * We already have a `tombstone_{warn,fail}_threshold` config items, inherited from C*, currently unused. Should we use these instead of adding a new item. My main problem with this is the name: `tombstone_fail_threshold` doesn't really tell a user that this just means the page is cut (it is accurate for unpaged queries though).

Certainly we should add a new item (if a new configuration is really warranted).