@denesb 2022-07-19
> > Fixes: #7689 Refs: #3672
> 
> Please also check if it fixes or somehow related to the following issues:
> 
>     * [Scan timeout after all data TTL exceeded #3914](https://github.com/scylladb/scylla/issues/3914) Scan timeout after all data TTL exceeded

Should fix this.

> 
>     * [Alternator's Scan should be ok with long strings of tombstones #7933](https://github.com/scylladb/scylla/issues/7933) Alternator's Scan should be ok with long strings of tombstones

Should fix this.

> 
>     * [Long reactor stalls when applying a large number of range tombstones to the cache #10282](https://github.com/scylladb/scylla/issues/10282) Long reactor stalls when applying a large number of range tombstones to the cache

Unrelated. This is related to the cache.

> 
> 
> If any issue is "related" but not fixes, please explain the nature of this relationship and why it doesn't fix - just saying "Refs" doesn't explain enough.