@denesb 2022-07-22
Force-pushed to save my progress so far. Not all comments are addressed yet, there will be a next version next week, nevertheless, just in case somebody wants to have a look:
New in v2:
* renamed config item query_tombstone_limit -> query_tombstone_page_limit
* use max-int for unlimited case and rename it to max to match partition and row limits
* move all users of read_command IDL constructor to the other, safer constructor
* have `storage_proxy::get_tombstone_limit()` return tagged int
* added handling of mismatched replica continue positions to digest resolver

Not addressed yet:
* count partition tombstones
* empty pages should be forwarded to client
* result_generator::_empty
* more tests