@scylladb-promoter 2022-07-22
CI state **`FAILURE`** - https://jenkins.scylladb.com/job/releng/job/Scylla-CI/1482/