@avikivity 2022-07-24
> * move all users of read_command IDL constructor to the other, safer constructor

You can extract this one too, if it saves you work later.