@denesb 2022-07-26
New in v3:
* Added comprehensive tests:
    - row tombstone prefix and span
    - range tombstone prefix and span
    - empty rows prefix and span
    - partition tombstones prefix and span
    - static row tombstones prefix and span
* Fixed issues found by tests, including a pre-existing obscure [IDL Frame truncated](https://github.com/scylladb/scylla/issues/9482) that they started to reliably reproduce.
* Implemented stopping on partition tombstones.
* Dropped internal client iterating on empty pages, this is left to the external client (driver), which is the whole point of this exercise: the empty pages act as a sort of heartbeat for a query that would otherwise time out.

TODO:
* multi-node sanity test
* multi-version sanity test

Removing the RFC tag, the branch is production ready pending the sanity tests I mentioned above.