@denesb 2022-07-26
> Does the cookie sent to the client get advanced when we send an empty page? If not, then if the 1865 cache misses, we'll keep looping on the same position.

Yes, the cookie sent to the client contains the last-position as provided by the replica. If the replica doesn't provide it that is a bug that will trigger an assert in the pager.