@avikivity 2022-07-26
Please run perf_simple_query before/after. I'm curious if the extra allocations I identified have an impact.