@denesb 2022-07-29
New in v4:
* Return `rpc::optional<full_position>` from `query_digest_locally()` to avoid an additional task to convert return values on the hot path.
* Only override last-position on query result if: empty-replica-pages feature is set and the min position is smaller than the current one.
* Split out to tangential bugfixes: https://github.com/scylladb/scylladb/pull/11139 and https://github.com/scylladb/scylladb/pull/11137. Both are merged.
* Avoid copy in `digest_read_resolver::digests_match()`.
* Added comment above `query_result_builder::_stop` explaining why we need to remember that we requested to stop.
* Adjusted config item reference in `query_result_builder::bump_and_check_tombstone_limit()`.

Change to query-pagers (removing `_row_count > 0` condition) is safe: there are not empty pages in an old or mixed cluster.

TODO:
* Sanity tests
* Perf tests