@denesb 2022-07-29
New in v5:
* Fix broken filtering queries.
* Remove incorrect assumption that last-position is always set in new cluster -- it is not set on truly empty pages (when there is no result at all).
* Added test for no-result queries.

TODO:
* More filtering+empty-pages tests.
* Multi-node sanity-test.
* Mixed cluster sanity-test.
* Perf test.