@denesb 2022-08-03
Upgrade sanity test found that I forgot to update `read_command::tombstone_limit = 0` -> `read_command::tombstone_limit = query::max_tombstones`. This generated a ton of read repairs in a mixed cluster. I guess this is why sanity tests are needed, even if the code was prepared to cope with mixed clusters.