@denesb 2022-08-03
I fixed the issue above. I'm still not done with my TODO list, I pushed so I can kick off a BYO build. I think for an intrusive change like this a simple CI run is not enough.