@denesb 2022-08-03
BYO job: https://jenkins.scylladb.com/job/scylla-master/job/byo/job/byo_build_tests_dtest/1496/