@scylladb-promoter 2022-08-03
CI state **`FAILURE`** - https://jenkins.scylladb.com/job/releng/job/Scylla-CI/1670/