@denesb 2022-08-04
New in v6 (compared to v5):
* rebased
* Fixed wrong default value in read_command IDL
* Fixed segfault in last position propagation
* completed sanity tests