@denesb 2022-08-04
Perf test: `build/release/test/perf/perf_simple_query -c1 -m2G --concurrency=100 --duration=5`
Before:
```
138843.29 tps ( 62.1 allocs/op,  12.1 tasks/op,   42359 insns/op,        0 errors)
142232.68 tps ( 62.1 allocs/op,  12.1 tasks/op,   42362 insns/op,        0 errors)
145798.97 tps ( 62.1 allocs/op,  12.1 tasks/op,   42363 insns/op,        0 errors)
141322.78 tps ( 62.1 allocs/op,  12.1 tasks/op,   42350 insns/op,        0 errors)
135312.45 tps ( 62.1 allocs/op,  12.1 tasks/op,   42365 insns/op,        0 errors)
```
After:
```
138843.29 tps ( 62.1 allocs/op,  12.1 tasks/op,   42359 insns/op,        0 errors)
142232.68 tps ( 62.1 allocs/op,  12.1 tasks/op,   42362 insns/op,        0 errors)
145798.97 tps ( 62.1 allocs/op,  12.1 tasks/op,   42363 insns/op,        0 errors)
141322.78 tps ( 62.1 allocs/op,  12.1 tasks/op,   42350 insns/op,        0 errors)
135312.45 tps ( 62.1 allocs/op,  12.1 tasks/op,   42365 insns/op,        0 errors)
```

No material difference.